refactor: 사용자 삭제 시 진행 중인 주문 여부 확인 로직 추가
